requirements.txt
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-docling==0.2.0
langchain-experimental==0.3.4
langchain-google-genai==2.0.8
langchain-huggingface==0.1.2
langchain-milvus==0.1.8
langchain-ollama==0.2.2
langchain-openai==0.3.0
langchain-postgres==0.0.12
langchain-text-splitters==0.3.5
sentence-transformers==2.6.1
transformers==4.43.1
bitsandbytes==0.42.0

torchvision
torchaudio
flask
python-dotenv==1.0.1
docx2txt==0.8
unstructured==0.16.5
unstructured-client==0.22.0
XlsxWriter==3.2.0
python-pptx>=1.0.2,<2.0.0
openpyxl>=3.1.5,<4.0.0
tqdm
milvus-lite==2.4.7
pymilvus==2.5.3
lxml==5.1.0
rank_bm25
openai==1.59.6
google-ai-generativelanguage==0.6.15
pillow==10.4.0
ragas==0.1.15
jq
flashrank
accelerate==0.34.0
pydantic==2.10.5
pydantic_core==2.27.2
psycopg2-binary==2.9.9
deepeval
nltk==3.9.1